John & Dom!!!
Still stuck on the "First" problem we where having in class...
Made a mistake and deleted something I wasn't supposed to and can't get
is back..
It should show on the pull request, don't accept it! IGetClaimInteractor
should interfer with it, so don't accept that.

Check the changes I made and make sure they are okay...
Push as soon as you can so I can pick up where I left off.

But John, help us with the "First" problem ~ Its in the DeviceViewModel